+++ src/index.ts
@@ -11,7 +11,6 @@ export {
 export { default as Card } from "./components/Card/Card";
 export { default as Checkbox } from "./components/Checkbox/Checkbox";
 export { default as Container } from "./components/Container/Container";
-export { default as DateRangeForm } from "./components/DateRangeForm/DateRangeForm";
 export { default as HeaderWithSearch } from "./components/RNHeaderWithSearch/RNHeaderWithSearch";
 export { default as Heading } from "./components/Heading/Heading";
 export { default as Hero } from "./components/Hero/Hero";